pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo-tester"
version = "0.1.0"
description = "GitHub repository safety scanner"
requires-python = ">=3.11"
dependencies = [
    "click==8.1.7",
    "requests==2.33.1",
    "python-Levenshtein==0.27.3",
]

[project.scripts]
repo-tester = "repo_tester.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/repo_tester"]
